Split giftcard mutations
