+++ saleor/graphql/giftcard/bulk_mutations/__init__.py
@@ -1,0 +1,11 @@
+from .gift_card_bulk_activate import GiftCardBulkActivate
+from .gift_card_bulk_create import GiftCardBulkCreate
+from .gift_card_bulk_deactivate import GiftCardBulkDeactivate
+from .gift_card_bulk_delete import GiftCardBulkDelete
+
+__all__ = [
+    "GiftCardBulkActivate",
+    "GiftCardBulkCreate",
+    "GiftCardBulkDeactivate",
+    "GiftCardBulkDelete",
+]
