+++ saleor/graphql/giftcard/bulk_mutations/gift_card_bulk_activate.py
@@ -1,0 +1,51 @@
+import graphene
+from django.core.exceptions import ValidationError
+
+from ....core.tracing import traced_atomic_transaction
+from ....giftcard import events, models
+from ....giftcard.error_codes import GiftCardErrorCode
+from ....giftcard.utils import is_gift_card_expired
+from ....permission.enums import GiftcardPermissions
+from ...app.dataloaders import get_app_promise
+from ...core import ResolveInfo
+from ...core.descriptions import ADDED_IN_31
+from ...core.mutations import BaseBulkMutation
+from ...core.types import GiftCardError, NonNullList
+from ...plugins.dataloaders import get_plugin_manager_promise
+from ..types import GiftCard
+
+
+class GiftCardBulkActivate(BaseBulkMutation):
+    class Arguments:
+        ids = NonNullList(
+            graphene.ID, required=True, description="List of gift card IDs to activate."
+        )
+
+    class Meta:
+        description = "Activate gift cards." + ADDED_IN_31
+        model = models.GiftCard
+        object_type = GiftCard
+        permissions = (GiftcardPermissions.MANAGE_GIFT_CARD,)
+        error_type_class = GiftCardError
+
+    @classmethod
+    def clean_instance(cls, _info: ResolveInfo, instance):
+        if is_gift_card_expired(instance):
+            raise ValidationError(
+                "Cannot activate expired card.",
+                code=GiftCardErrorCode.EXPIRED_GIFT_CARD.value,
+            )
+
+    @classmethod
+    @traced_atomic_transaction()
+    def bulk_action(cls, info: ResolveInfo, queryset, /):
+        queryset = queryset.filter(is_active=False)
+        gift_card_ids = [gift_card.id for gift_card in queryset]
+        app = get_app_promise(info.context).get()
+        queryset.update(is_active=True)
+        events.gift_cards_activated_event(
+            gift_card_ids, user=info.context.user, app=app
+        )
+        manager = get_plugin_manager_promise(info.context).get()
+        for card in models.GiftCard.objects.filter(id__in=gift_card_ids):
+            manager.gift_card_status_changed(card)
