+++ saleor/graphql/giftcard/mutations/__init__.py
@@ -1,0 +1,17 @@
+from .gift_card_activate import GiftCardActivate
+from .gift_card_add_note import GiftCardAddNote
+from .gift_card_create import GiftCardCreate
+from .gift_card_deactivate import GiftCardDeactivate
+from .gift_card_delete import GiftCardDelete
+from .gift_card_resend import GiftCardResend
+from .gift_card_update import GiftCardUpdate
+
+__all__ = [
+    "GiftCardActivate",
+    "GiftCardAddNote",
+    "GiftCardCreate",
+    "GiftCardDeactivate",
+    "GiftCardDelete",
+    "GiftCardResend",
+    "GiftCardUpdate",
+]
