+++ saleor/graphql/giftcard/tests/mutations/test_gift_card_create.py
@@ -90,7 +90,9 @@ CREATE_GIFT_CARD_MUTATION = """
 """
 
 
-@mock.patch("saleor.graphql.giftcard.mutations.send_gift_card_notification")
+@mock.patch(
+    "saleor.graphql.giftcard.mutations.gift_card_create.send_gift_card_notification"
+)
 def test_create_never_expiry_gift_card(
     send_notification_mock,
     staff_api_client,
@@ -185,7 +187,9 @@ def test_create_never_expiry_gift_card(
     )
 
 
-@mock.patch("saleor.graphql.giftcard.mutations.send_gift_card_notification")
+@mock.patch(
+    "saleor.graphql.giftcard.mutations.gift_card_create.send_gift_card_notification"
+)
 def test_create_gift_card_by_app(
     send_notification_mock,
     app_api_client,
@@ -501,7 +505,9 @@ def test_create_gift_card_with_zero_balance_amount(
     assert errors[0]["code"] == GiftCardErrorCode.INVALID.name
 
 
-@mock.patch("saleor.graphql.giftcard.mutations.send_gift_card_notification")
+@mock.patch(
+    "saleor.graphql.giftcard.mutations.gift_card_create.send_gift_card_notification"
+)
 def test_create_gift_card_with_expiry_date(
     send_notification_mock,
     staff_api_client,
